Remove unnecessary .take() calls in Rust
These are not necessary because the compiler can statically reason about temporary "uninitialized" variables. Unfortunately this doesn't work for self.root in the Tree methods, as there the uninitialized memory could be perceivable when those methods panic.

Looking at the assembly, this removes some unnecessary drop calls, so this should result in a slight speedup.